Fixed examples after refactor
