+++ examples/chicken-cli/checkin-cli.go
